requirements.txt
streamlit
torch
torchvision
pillow
numpy
ftfy
regex
git+https://github.com/openai/CLIP.git